refactor: move YamlRendererBase to contrib/yaml/
YamlRendererBase is a concrete renderer like Html/Svg/Markdown — it
belongs in contrib/, not in a top-level renderers/ package.
Remove the now-empty renderers/ package.
