+++ src/genro_builders/__init__.py
@@ -37,11 +37,11 @@ from genro_builders.builder_bag import BuilderBag, BuilderBagNode, Component
 from genro_builders.compiler import BagCompilerBase, compiler
 from genro_builders.component_proxy import ComponentProxy
 from genro_builders.component_resolver import ComponentResolver
+from genro_builders.contrib.yaml import YamlRendererBase
 from genro_builders.manager import BuilderManager
 from genro_builders.pointer import PointerInfo, is_pointer, parse_pointer, scan_for_pointers
 from genro_builders.reactive_manager import ReactiveManager
 from genro_builders.renderer import BagRendererBase, RenderNode, renderer
-from genro_builders.renderers import YamlRendererBase
 
 __version__ = "0.16.0"
