+++ src/genro_builders/contrib/__init__.py
@@ -1,12 +1,13 @@
 # Copyright 2025 Softwell S.r.l. - SPDX-License-Identifier: Apache-2.0
-"""Contributed builders — ready-to-use domain-specific builders.
+"""Contributed builders and renderers — ready-to-use domain-specific components.
 
-These builders are shipped with genro-builders but are not part of the
-core framework. They are optional and only loaded when explicitly imported.
+These are shipped with genro-builders but are not part of the core framework.
+They are optional and only loaded when explicitly imported.
 
-Available builders:
-    - ``genro_builders.contrib.html`` — HtmlBuilder (W3C HTML5, 112 elements)
-    - ``genro_builders.contrib.markdown`` — MarkdownBuilder (headings, lists, tables, code)
-    - ``genro_builders.contrib.svg`` — SvgBuilder (W3C SVG, shapes, text, gradients, filters)
-    - ``genro_builders.contrib.xsd`` — XsdBuilder (schema-driven XML from XSD files)
+Available:
+    - ``genro_builders.contrib.html`` — HtmlBuilder, HtmlRenderer (W3C HTML5)
+    - ``genro_builders.contrib.markdown`` — MarkdownBuilder, MarkdownRenderer
+    - ``genro_builders.contrib.svg`` — SvgBuilder, SvgRenderer (W3C SVG)
+    - ``genro_builders.contrib.xsd`` — XsdBuilder (schema-driven XML from XSD)
+    - ``genro_builders.contrib.yaml`` — YamlRendererBase (YAML output)
 """
