+++ src/genro_builders/contrib/yaml/__init__.py
@@ -1,0 +1,6 @@
+# Copyright 2025 Softwell S.r.l. - SPDX-License-Identifier: Apache-2.0
+"""YAML renderer for genro-builders."""
+
+from genro_builders.contrib.yaml.yaml_renderer import YamlRendererBase
+
+__all__ = ["YamlRendererBase"]
